+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Logs
 
+## 0.0.14
+   1. Added get curated photos
+   2. Documentation updated for photo and video. For more visit [Documentation](https://pub.dev/documentation/pexels_photos_videos/latest/)
+
 ## 0.0.13
 
 * Major update
    1. Added get video details
-   2. Change in PexelsMedia class now no need pass authorization key again and and again.
-   3. Documentation updated for photo and video. For more visit [Documentation](https://pub.dev/documentation/pexels_photos_videos/latest/)
+   2. Change in PexelsMedia class and now no need to pass authorization key again and and again.
+   3. Documentation updated for photo and video. For more visit [Documentation](https://pub.dev/documentation/pexels_photos_videos/0.0.13/)
 
 ## 0.0.12